fara complicatie

+++ java/Lab1.java
@@ -365,85 +365,23 @@ public class Lab1 {
         System.out.println("\nNumar total de profi creati: " + Prof.getProfesori());
     
 
+        // 8. Scriere in fisiere
         File folder = new File(".");
         File[] files = folder.listFiles((dir, name) -> name.endsWith(".txt") && !name.equals("profText.txt"));
 
-        for(int i = 0; i<files.length; i++){
-            files[i].delete();
-        }
-
-
-
-
-        String numeScrise[] = new String[n];
-        int dupes = 0;
-        int nrNumescrise = 0;
-        
-        
-
-        for(int i = 0; i < n; i++){
-            boolean isDupe = false;
-
-            for(int j = 0; j < n; j++){
-                if(numeScrise[j] != null){
-                    nrNumescrise++;
-                }
+        if (files != null) {
+            System.out.println("\n--- Ștergerea fișierelor .txt existente (excluzând profText.txt) ---");
+            
+            for(int i = 0; i<files.length; i++){
+                System.out.println("Șterg: " + files[i].getName());
+                files[i].delete();
             }
-
-            for(int l = 0; l < nrNumescrise; l++){
-                if (univer[i].getNume().equals(numeScrise[l])){
-                    isDupe = true;
-                }
-            }
-
-            if(!isDupe){
-                try {
-                    numeScrise[i] = univer[i].getNume();
-                    FileWriter fw = new FileWriter(univer[i].getNume() + ".txt");
-                    fw.write(univer[i].getNume() + "\n");
-                    fw.write(univer[i].getObiect() + "\n");
-                    fw.write(univer[i].getRautaciune() + "\n");
-                    fw.write(univer[i].getGalagie() + "\n");
-                    fw.write(univer[i].getNrGrupe() + "\n");
-                    for(int j = 0; j < univer[i].getNrGrupe(); j++){    
-                        fw.write(univer[i].getNrStudenti()[j] + " ");
-                    }
-
-                    fw.close();
-                } catch (Exception e) {
-                    System.out.println("error");
-                }
-
-            }else{
-                System.out.println("nume duplicat, creeam fisier cu suffix (n)");
-
-                try {
-                    numeScrise[i] = univer[i].getNume();
-                    FileWriter fw = new FileWriter(univer[i].getNume() + "" + ".txt");
-                    fw.write(univer[i].getNume() + "\n");
-                    fw.write(univer[i].getObiect() + "\n");
-                    fw.write(univer[i].getRautaciune() + "\n");
-                    fw.write(univer[i].getGalagie() + "\n");
-                    fw.write(univer[i].getNrGrupe() + "\n");
-                    for(int j = 0; j < univer[i].getNrGrupe(); j++){    
-                        fw.write(univer[i].getNrStudenti()[j] + " ");
-                    }
-
-                    fw.close();
-                } catch (Exception e) {
-                    System.out.println("error");
-                }
-            }
-
 
         }
-
-
-/*            if(univer[i] != null && "add logic"){
-
-                
+        for(int i = 0; i< univer.length; i++){
+            
+           if(univer[i] != null){
                 try {
-                    numeScrise[i] = univer[i].getNume();
                     FileWriter fw = new FileWriter(univer[i].getNume() + ".txt");
                     fw.write(univer[i].getNume() + "\n");
                     fw.write(univer[i].getObiect() + "\n");
@@ -459,26 +397,9 @@ public class Lab1 {
                     System.out.println("error");
                 }
             }else{
-                System.out.println("name found twice! writing as Nume + (i)");
-                dupes++;
-                try {
-                    numeScrise[i] = univer[i].getNume() + "(" + dupes + ")";
-                    FileWriter fw = new FileWriter(univer[i].getNume() +  "(" + dupes + ")" + ".txt");
-                    fw.write(univer[i].getNume() + "\n");
-                    fw.write(univer[i].getObiect() + "\n");
-                    fw.write(univer[i].getRautaciune() + "\n");
-                    fw.write(univer[i].getGalagie() + "\n");
-                    fw.write(univer[i].getNrGrupe() + "\n");
-                    for(int j = 0; j < univer[i].getNrGrupe(); j++){    
-                        fw.write(univer[i].getNrStudenti()[j] + " ");
-                    }
-
-                    fw.close();
-                } catch (Exception e) {
-                    System.out.println("error");
-                }
+                System.out.println("Found a null entry, skipping.");
             }
-         */
+        
         
           
     
@@ -488,6 +409,6 @@ public class Lab1 {
     }
 
 
-    
+    }
 
 }
